Agent/requirements.txt
fastapi
uvicorn
smolagents==1.13.0
requests
python-dotenv
pyyaml
ddgs